requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.6.3
black==24.10.0
cachetools==4.2.2
certifi==2024.8.30
click==8.1.7
exceptiongroup==1.2.2
gigachat==0.1.35
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-telegram-bot==13.13
pytz==2024.2
six==1.16.0
sniffio==1.3.1
tomli==2.0.2
tornado==6.4.1
typing_extensions==4.12.2
tzlocal==5.2
